Schema generator improvements:
    * More granular documetation
    * Convinience method with single descriptor
    * Alll the implementation is encapsulated into the object body
